ready work

+++ src/controllers/notesController.js
@@ -1,15 +1,12 @@
 import createHttpError from "http-errors";
 import { Note } from "../models/note.js";
 
-
 export const getAllNotes = async (req, res, next) => {
   try {
     const { page = 1, perPage = 10, tag, search } = req.query;
-
     const skip = (Number(page) - 1) * Number(perPage);
 
-    const query = Note.find();
-
+    const query = Note.find().where("userId").equals(req.user._id);
     if (tag) query.where("tag").equals(tag);
     if (search) query.where({ $text: { $search: search } });
 
@@ -28,18 +25,11 @@ export const getAllNotes = async (req, res, next) => {
   }
 };
 
-export const getNoteById  = async (req, res, next) => {
-  try{
-
-    const {noteId} = req.params;
-    const note = await Note.findById(noteId);
-
-    if(!note){
-      throw createHttpError(404, 'Note not found');
-
-    }
-
-    res.status(200).json((note));
+export const getNoteById = async (req, res, next) => {
+  try {
+    const note = await Note.findOne({ _id: req.params.noteId, userId: req.user._id });
+    if (!note) throw createHttpError(404, "Note not found");
+    res.status(200).json(note);
   } catch (error) {
     next(error);
   }
@@ -47,7 +37,7 @@ export const getNoteById  = async (req, res, next) => {
 
 export const createNote = async (req, res, next) => {
   try {
-    const note = await Note.create(req.body);
+    const note = await Note.create({ ...req.body, userId: req.user._id });
     res.status(201).json(note);
   } catch (error) {
     next(error);
@@ -55,37 +45,25 @@ export const createNote = async (req, res, next) => {
 };
 
 export const deleteNote = async (req, res, next) => {
-  try{
-    const {noteId} = req.params;
-    const note = await Note.findOneAndDelete({_id: noteId});
-    if(!note){
-      throw createHttpError(404, "Note not found");
-
-    }
-
+  try {
+    const note = await Note.findOneAndDelete({ _id: req.params.noteId, userId: req.user._id });
+    if (!note) throw createHttpError(404, "Note not found");
     res.status(200).json(note);
-
-  } catch (error){
+  } catch (error) {
     next(error);
   }
-
 };
 
 export const updateNote = async (req, res, next) => {
-    try{
-    const {noteId} = req.params;
+  try {
     const note = await Note.findOneAndUpdate(
-    {_id: noteId},
-    req.body,
-    { new: true },
+      { _id: req.params.noteId, userId: req.user._id },
+      req.body,
+      { new: true }
     );
-
-    if(!note){
-      throw createHttpError(404, "Note not found");
-
-    }
+    if (!note) throw createHttpError(404, "Note not found");
     res.status(200).json(note);
-  } catch (error){
-      next(error);
-    }
+  } catch (error) {
+    next(error);
+  }
 };
